requirements.txt
python-dotenv
langchain==0.3.4
langchain-openai==0.2.3
Flask==3.0.3
openai==1.52.2
langchain-community==0.3.3
pydantic==2.9.2
faiss-cpu==1.9.0
langchain-chroma==0.1.4
chromadb==0.5.15
unstructured==0.16.2
streamlit==1.39.0
streamlit-chat==0.1.1
pymongo==4.10.1
